Add support for building images with podman
Podman is a daemonless, open-source tool for managing containers, images,
volumes, and pods, offering a user-friendly alternative to Docker.

As a user of Fedora Sway atomic, it would be great if it was possible to
build armbian with podman. This is because all immutable fedora
distributions have podman pre-installed. Additionally, installing docker
properly on fedora based distributions has been a bit of hit and miss
for me (but that just might be a me problem).

Currently, building with podman requires rootful container. This
probably is not any worse than privileged docker, but I am not an
expert.

It is possible to execute the docker container builds as normal user,
but that would require exporting and re-importing the container to the
root user anyway, so I am just doing the container build as root user
for now.

Additionally, some lables are required for mounts since by default (even
for privileged containers), podman adds nodev and nosuid, which is
will cause the build to fail.

Signed-off-by: Ayush Singh <ayush@beagleboard.org>